extract splash screen
