+++ crates/client/src/app/input/mod.rs
@@ -5,3 +5,4 @@ pub mod hotbar_assignment;
 pub mod inventory;
 pub mod multi_action;
 pub mod npc_interaction;
+pub mod splash;
